chore: 🤖 reorganize tests
